doa: ignoring kill_process() if already terminated

+++ rope/base/oi/doa.py
@@ -77,6 +77,8 @@ class PythonFileRunner(object):
 
     def kill_process(self):
         """Stop the process"""
+        if self.process.returncode is not None:
+            return
         try:
             if hasattr(self.process, 'terminate'):
                 self.process.terminate()
